Fix screen size, bump min tcod version.
The window was rendering way too small at my native desktop resolution
(5760x3240). It now defaults to 2560x1440. This is probably too large
for a lot of contexts. The window will scale properly if you manually
resize it. Otherwise just change the values for screen_width and
screen_height in main.py.